server/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
astroid==3.2.4
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
dill==0.3.8
Django==5.1.1
djangorestframework==3.15.2
frozenlist==1.4.1
idna==3.8
isort==5.13.2
mccabe==0.7.0
multidict==6.0.5
platformdirs==4.2.2
psycopg2==2.9.9
PyJWT==2.9.0
pylint==3.2.7
requests==2.32.3
sqlparse==0.5.1
tomlkit==0.13.2
twilio==9.2.4
urllib3==2.2.2
yarl==1.9.8
